requirements.txt
Flask==3.0.3
google-api-python-client==2.127.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
gunicorn==22.0.0
openpyxl==3.1.2
pandas==2.2.2
